+++ oci_api/image/repository.py
@@ -16,10 +16,11 @@ import json
 import pathlib
 import logging
 from oci_spec.image.v1 import Index, ImageLayout
+from oci_api.util import id_to_digest, digest_to_id
 from oci_api import oci_config, OCIError
-from oci_api.util import id_to_digest
 from oci_api.util.file import rm
 from .image import Image
+from .exceptions import ImageExistsException
 
 log = logging.getLogger(__name__)
 
@@ -29,14 +30,15 @@ class Repository():
         self.name = name
         self.index = None
         self.images = {}
-        try:
+        repositories_path = pathlib.Path(oci_config['global']['path'], 'repositories')
+        index_file_path = repositories_path.joinpath(self.name + '.json')
+        if index_file_path.is_file():
             self.load()
-        except:
-            pass
 
     def load(self):
         repositories_path = pathlib.Path(oci_config['global']['path'], 'repositories')
         index_file_path = repositories_path.joinpath(self.name + '.json')
+        log.debug('Start loading index file (%s)' % index_file_path)
         self.index = Index.from_file(index_file_path)
         for manifest_descriptor in self.index.get('Manifests'):
             manifest_annotations = manifest_descriptor.get('Annotations')
@@ -47,6 +49,7 @@ class Repository():
             image.load(manifest_id)
             if image is not None:
                 self.images[image.tag] = image
+        log.debug('Finish loading index file (%s)' % index_file_path)
 
     def save(self):
         if self.index is None:
@@ -56,10 +59,6 @@ class Repository():
             repositories_path.mkdir(parents=True)
         index_file_path = repositories_path.joinpath(self.name + '.json')
         self.index.save(index_file_path)
-        oci_layout_file_path = repositories_path.joinpath('oci-layout')
-        if not oci_layout_file_path.is_file():
-            oci_layout = ImageLayout(version='1.0.0')
-            oci_layout.save(oci_layout_file_path)
 
     def remove(self):
         if len(self.images) != 0:
@@ -77,12 +76,13 @@ class Repository():
         except:
             raise OCIError('Image (%s) does not exist in this repository' % tag)
 
-    def create_image(self, tag, rootfs_tar_file, image_config):
+    def import_image(self, tag, rootfs_tar_path, image_config):
         image = self.images.get(tag, None)
         if image is not None:
-            raise OCIError('Image tag (%s) already exist' % tag)
+            raise ImageExistsException('Image tag (%s) already exist' % tag)
         image = Image(self.name, tag)
-        manifest_descriptor = image.create(rootfs_tar_file, image_config)
+        manifest_descriptor = image.create(rootfs_tar_path, image_config)
+        manifest_descriptor.add('Annotations', {'org.opencontainers.image.ref.name': tag})
         self.images[tag] = image
         if self.index is None:
             self.index = Index(
@@ -110,3 +110,55 @@ class Repository():
             self.remove()
             return
         self.save()
+
+    def save_image(self, tag, repository_layout_path):
+        oci_layout = ImageLayout(version='1.0.0')
+        oci_layout_file_path = repository_layout_path.joinpath('oci-layout')
+        log.debug('Creating oci layout file (%s)' % oci_layout_file_path)
+        oci_layout.save(oci_layout_file_path)
+        image = self.get_image(tag)
+        manifest_digest = id_to_digest(image.id)
+        for manifest_descriptor in self.index.get('Manifests'):
+            if manifest_descriptor.get('Digest') == manifest_digest:
+                index = Index(manifests=[manifest_descriptor])
+                index_file_path = repository_layout_path.joinpath('index.json')
+                log.debug('Creating index file (%s)' % index_file_path)
+                index.save(index_file_path)
+                image.export(repository_layout_path)
+                return
+
+    def load_image(self, tag, repository_layout_path):
+        if tag in self.images:
+            raise ImageExistsException('Image tag (%s) already exist' % tag)
+
+        oci_file_path = repository_layout_path.joinpath('oci-layout')
+        if not oci_file_path.is_file():
+            raise OCIError('There is no oci-layout file in (%s)' % str(repository_layout_path))
+        oci_layout = ImageLayout.from_file(oci_file_path)
+
+        index_file_path = repository_layout_path.joinpath('index.json')
+        if not index_file_path.is_file():
+            raise OCIError('There is no index.json file in (%s)' % str(repository_layout_path))
+        index = Index.from_file(index_file_path)
+        manifest_descriptors = index.get('Manifests')
+        if len(manifest_descriptors) == 0:
+            raise OCIError('Index (%s) has no manifest' % str(index_file_path))
+        if len(manifest_descriptors) > 1:
+            raise OCIError('Index (%s) has (%d) manifests, only one supported' % 
+            (str(index_file_path), len(manifest_descriptors)))
+        manifest_descriptor = manifest_descriptors[0]
+        manifest_digest = manifest_descriptor.get('Digest')
+        manifest_id = digest_to_id(manifest_digest)
+        manifest_descriptor.add('Annotations', {'org.opencontainers.image.ref.name': tag})
+
+        image = Image(self.name, tag)
+        image.copy(manifest_id, repository_layout_path)
+        self.images[tag] = image
+
+        if self.index is None:
+            self.index = index
+        else: 
+            manifests = self.index.get('Manifests')
+            manifests.append(manifest_descriptor)
+        self.save()
+        return image